Move connect and scan into separate packages
