+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/CardReaderModule.kt
@@ -3,6 +3,10 @@ package com.woocommerce.android.ui.prefs.cardreader
 import com.woocommerce.android.di.FragmentScope
 import com.woocommerce.android.ui.prefs.cardreader.CardReaderModule.CardReaderConnectFragmentModule
 import com.woocommerce.android.ui.prefs.cardreader.CardReaderModule.CardReaderScanFragmentModule
+import com.woocommerce.android.ui.prefs.cardreader.connect.CardReaderConnectFragment
+import com.woocommerce.android.ui.prefs.cardreader.connect.CardReaderConnectModule
+import com.woocommerce.android.ui.prefs.cardreader.scan.CardReaderScanFragment
+import com.woocommerce.android.ui.prefs.cardreader.scan.CardReaderScanModule
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
