+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/scan/CardReaderScanViewModel.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.prefs.cardreader
+package com.woocommerce.android.ui.prefs.cardreader.scan
 
 import com.woocommerce.android.di.ViewModelAssistedFactory
 import com.woocommerce.android.util.CoroutineDispatchers
